fix: address Phase 3 code review minor issues
- MINOR ISSUE 1: Remove unnecessary 'as u32' cast at device.rs:185 and device.rs:490 — count is already u32 from desc.len/4096 division
- MINOR ISSUE 2: Replace manual (num_pages + 63) / 64 with num_pages.div_ceil(64) in reclaimed_bitmap.rs:32
- MINOR ISSUE 3: Remove unused pub(crate) use re-export and #[allow(unused_imports)] from mod.rs:7-8

All balloon tests pass (34/34).

+++ src/devices/src/virtio/balloon/device.rs
@@ -182,7 +182,7 @@ impl Balloon {
                         "balloon: marking FRQ pages as reported-free: start_pfn={:#x} count={}",
                         start_pfn, count
                     );
-                    bitmap.mark_range(start_pfn, count as u32);
+                    bitmap.mark_range(start_pfn, count);
                 }
             }
 
@@ -487,7 +487,7 @@ impl Balloon {
                                 "balloon: marking PHQ pages as reported-free: start_pfn={:#x} count={}",
                                 start_pfn, count
                             );
-                            bitmap.mark_range(start_pfn, count as u32);
+                            bitmap.mark_range(start_pfn, count);
                         }
                     }
                 }
